+++ js/phpSysInfo/phpsysinfo.js
@@ -1431,35 +1431,21 @@ function settimer(xml) {
 }
 
 $(document).ready(function buildpage() {
-    var i = 0, cookie_language = null, cookie_template = null;
+    var i = 0, cookie_template = null, cookie_language = null;
 
-    if ($("#template option").size() < 2) {
-        switchStyle($("#template").val().toString());
-    } else {
-        $('#template').show();
-        $('span[id=lang_044]').show();
-        cookie_template = readCookie("template");
-        if (cookie_template !== null) {
-            $("#template").val(cookie_template);
-        }
-        $("#template").change(function changeTemplate() {
-            switchStyle($("#template").val().toString());
-            createCookie('template', $("#template").val().toString(), 365);
-            return false;
-        });
-        switchStyle($("#template").val().toString());
-    }
-    if ($("#lang option").size() < 2) {
-        current_language = $("#lang").val().toString();
+    if ($("#language option").size() < 2) {
+        current_language = $("#language").val().toString();
         changeLanguage();
         for (i = 0; i < plugin_liste.length; i += 1) {
             changeLanguage(plugin_liste[i]);
         }
     } else {
-        cookie_language = readCookie("language");
+        cookie_language = readCookie("psi_language");
         if (cookie_language !== null) {
             current_language = cookie_language;
-            $("#lang").val(current_language);
+            $("#language").val(current_language);
+        } else {
+            current_language = $("#language").val().toString();
         }
         changeLanguage();
         for (i = 0; i < plugin_liste.length; i += 1) {
@@ -1467,14 +1453,30 @@ $(document).ready(function buildpage() {
         }
         $('#lang').show();
         $('span[id=lang_045]').show();
-        $("#lang").change(function changeLang() {
+        $("#language").change(function changeLang() {
             var i = 0;
-            current_language = $("#lang").val().toString();
-            createCookie('language', current_language, 365);
+            current_language = $("#language").val().toString();
+            createCookie('psi_language', current_language, 365);
             changeLanguage();
             for (i = 0; i < plugin_liste.length; i += 1) {
                 changeLanguage(plugin_liste[i]);
             }
+            return false;
+        });
+    }
+    if ($("#template option").size() < 2) {
+        switchStyle($("#template").val().toString());
+    } else {
+        cookie_template = readCookie("psi_template");
+        if (cookie_template !== null) {
+            $("#template").val(cookie_template);
+        }
+        switchStyle($("#template").val().toString());
+        $('#template').show();
+        $('span[id=lang_044]').show();
+        $("#template").change(function changeTemplate() {
+            switchStyle($("#template").val().toString());
+            createCookie('psi_template', $("#template").val().toString(), 365);
             return false;
         });
     }