+++ CHANGELOG.md
@@ -7,7 +7,7 @@ phpSysInfo 3.2.4
 ----------------
  - [ADD] Solus to detected distros
 
- - [NEW] Language selection also for frontend "bootstrap"
+ - [NEW] language and template selection also for frontend "bootstrap"
 
 phpSysInfo 3.2.3
 ----------------